fixed slider issues
